Reuse input across runs on AWS
* Move AWS credentials from file to environment variable.
* Provide options to run `run-on-ec2.py` in setup only, skip setup and
full mode. This allows the user to reuse the setup files which were
uploaded for a larger k for any consecutive runs with a smaller k.

+++ aws/s3Manager.py
@@ -41,11 +41,12 @@ class S3Manager(object):
             Prefix=self.prefix
         )
         keysToDelete = []
-        for obj in response['Contents']:
-            keysToDelete.append({'Key': obj["Key"]})
-        self.s3Client.delete_objects(
-            Bucket=self.bucket,
-            Delete={
-                'Objects': keysToDelete
-            }
-        )
+        if 'Contents' in response:
+            for obj in response['Contents']:
+                keysToDelete.append({'Key': obj["Key"]})
+            self.s3Client.delete_objects(
+                Bucket=self.bucket,
+                Delete={
+                    'Objects': keysToDelete
+                }
+            )
